Work on Retrofit Query for recipes
Working on the query I encountered a problem. My Recipe API is limited
to only looking for q query and not cuisinetype and diettype. I'm going
to either need to upgrade my API or look for other one. For now I'm
going to focus on the user side of the program